my_django_project/requirements.txt
Django==4.1.1
asgiref==3.6.0
sqlparse==0.4.4
psycopg2-binary==2.9.6
Pillow==9.5.0
python-gettext==5.0
python-slugify==8.0.1
requests==2.30.0
passlib==1.7.4
PyJWT~=2.3.0
djangorestframework==3.14.0
Pygments==2.16.1